pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledgeforge"
version = "0.1.0"
description = "Universal RAG Knowledge System - shared brain for AI agents"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "markdown-it-py>=3.0.0",
    "python-frontmatter>=1.0.0",
    "watchdog>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "typer>=0.9.0",
    "pydantic-settings>=2.0.0",
    "mcp[cli]>=1.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-c>=0.21.0",
    "tree-sitter-cpp>=0.22.0",
    "tree-sitter-bash>=0.21.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
knowledgeforge = "knowledgeforge.interfaces.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
